+++ changelog.md
@@ -1,2 +1,6 @@
 # 1.0.0
-- Edit this file to change your mod's changelog.
+- Initial release
+
+# 1.0.1
+- Bugfixing for level end and restart from endscreen
+- Removing debug logs